refac: use component for next on/off times
