+++ src/routes/monitor/+page.svelte
@@ -6,19 +6,8 @@
     import MonitorChart from '$lib/components/monitorChart.svelte';
     import configData from '$lib/data/Configuration-Data.json';
     import { LatLng } from 'leaflet';
+    import NextOnOffTime from './components/nextOnOffTime.svelte';
 
-    // Day number of today
-    const getDayOfYear = (date: Date): number =>
-        Math.floor(
-            (date.getTime() - new Date(date.getFullYear(), 0, 0).getTime()) / (1000 * 60 * 60 * 24),
-        );
-    const dayOfYearForArray = getDayOfYear(new Date()) - 1;
-
-    // Today on / off
-    const todayOnHour = configData.computed_timers[dayOfYearForArray].on_time.hour;
-    const todayOnMinute = configData.computed_timers[dayOfYearForArray].on_time.minute;
-    const todayOffHour = configData.computed_timers[dayOfYearForArray].off_time.hour;
-    const todayOffMinute = configData.computed_timers[dayOfYearForArray].off_time.minute;
 
     // Time zone
     const timezone = configData.timezone;
@@ -61,12 +50,7 @@
 
             <div class="flex flex-col">
                 <div class="flex flex-col">
-                    <p class="text-lg font-semibold">
-                        Einschaltung heute: <span class="font-normal">{todayOnHour}:{todayOnMinute} Uhr</span>
-                    </p>
-                    <p class="text-lg font-semibold">
-                        Ausschaltung heute: <span class="font-normal">{todayOffHour}:{todayOffMinute} Uhr</span>
-                    </p>
+                    <NextOnOffTime />
                     <h2 class="mb-2 mt-5 text-center text-xl font-semibold">Terrarium Standort</h2>
                 </div>
                 <div>
